페이지 메뉴의 최대크기 지정, 페이지 이동시 게시판정보가 누락되지 않도록

+++ src/main/java/com/kpk/exam/demo/controller/UsrArticleController.java
@@ -73,7 +73,7 @@ public class UsrArticleController {
 		// 한 페이지당 10개씩
 		// 글 20개 --> 2
 		// 글 24개 --> 3
-		int pagesCount = (int) Math.ceil(articlesCount / itemsInAPage);
+		int pagesCount = (int) Math.ceil((double)articlesCount / itemsInAPage);
 		
 		List<Article> articles = articleService.getForPrintArticles(rq.getLoginedMemberId(), boardId, itemsInAPage, page);
